[BI-2055] renamed annotation from ExperimentSecured -> ExperimentCollaboratorSecured
